parser-ocr/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
PyMuPDF==1.25.2
pytesseract==0.3.13
Pillow==11.1.0
python-docx==1.1.2
